# Optimize the select_best_admin search for babelfish