### Description
In Major version 16 Babelfish introduced a new role bbf_admin_role which is a member of all the roles which created implicitly from TSQL. This increases the Load on any Grant/Revoke operations that happen since in Vanilla PG 16 the path to select the best admin is inefficient and loops over members of members to find the best admin and it is uniquely exacerbated by babelfish due to this role.
This commit optimizes the selection of best admin for babelfish by escaping the search for any TSQL roles conditionally.
Without the  changes on github the time taken to create database increases drastically and linearly:
```
2nd 'create db' stmt 4 ms

2000th 'create db' stmt 1024 ms

The average time consumed by eached create db stmt: 1651.895 ms
```
On Github with these changes we see the time taken to create database increasing linearly but not as drastically as it was before the changes:
```
2nd 'create db' stmt 5 ms

2000th 'create db' stmt 5089 ms

The average time consumed by eached create db stmt: 361.684 ms
```
So the speedup to create 2000 dbs achieved with this change is: (1651.895 - 361.684) * 2000
Which is significant.


### Issues Resolved

BABEL-4849

Authored-by: Kushaal Shroff [kushaal@amazon.com](mailto:kushaal@amazon.com)
Signed-off-by: Kushaal Shroff [kushaal@amazon.com](mailto:kushaal@amazon.com)

### Check List
- [x] Commits are signed per the DCO using --signoff 

By submitting this pull request, I confirm that my contribution is under the terms of the Apache 2.0 and PostgreSQL licenses, and grant any person obtaining a copy of the contribution permission to relicense all or a portion of my contribution to the PostgreSQL License solely to contribute all or a portion of my contribution to the PostgreSQL open source project.

For more information on following Developer Certificate of Origin and signing off your commits, please check [here](https://github.com/babelfish-for-postgresql/babelfish_extensions/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).